refactor use core/swap.hpp over swap.hpp
